preset list function now returns a json containing the preset name and corresponding image path
